cleaned up getting full trace manifest
